generateRoot: createPTR generates it's own root div, user passes in the container into which it should be placed
