+++ index.html
@@ -22,10 +22,8 @@
         gap: 4rem;
       }
 
-      #root {
-        display: flex;
-        justify-content: center;
-        box-sizing: border-box;
+      #root > div {
+        width: 100%;
       }
 
       .buttonGroup {
